frontend/requirements.txt
dash==2.8.0
gunicorn
plotly
pandas
dash_bootstrap_components
dash_renderjson
requests
pillow
numpy